requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

tree-sitter==0.20.4
tree-sitter-languages==1.10.2

faiss-cpu>=1.8.0
numpy>=1.24.3

watchdog==3.0.0

pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.1.1

cachetools==5.3.2

httpx==0.25.2
aiohttp==3.9.1

groq==0.4.1

pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0